add results from pairing

+++ AndroidTestAutomationStarter/app/src/main/java/com/novoda/androidstoreexample/models/Order.kt
@@ -1,5 +1,5 @@
 package com.novoda.androidstoreexample.models
 
 data class Order(val product: Product, val numberOfItems: Int) {
-    val amount: Int = product.price * numberOfItems
+    val amount: Int = product.price.toInt() * numberOfItems
 }
